pyproject.toml
[tool.poetry]
name = "finsy"
version = "0.1.0"
description = ""
authors = ["Bill Fisher <william.w.fisher@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.47.0"
macaddress = "^1.2.0"
parsy = "^1.4.0"
pyee = "^9.0.4"
pylev = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.3.0"
pytest-asyncio = "^0.18.1"
github-clone = "^1.2.0"
grpcio-tools = "^1.44.0"
mypy-protobuf = "^3.2.0"
mypy = "^0.961"
pdoc = "^12.0.2"
pyright = "^1.1.243"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = '--log-cli-format="%(created).03f %(levelname)s %(name)s %(message)s" --log-level=INFO --log-file-format="%(created).03f %(levelname)s %(name)s %(message)s"'

[tool.pyright]
ignore = ["finsy/proto/**/*_pb2.py"]

[tool.black]
target-version = ['py310']
required-version = "22.3.0"
extend-exclude = "_pb2(_grpc)?\\.pyi?$"
